requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
supabase>=2.3.0
psycopg2-binary>=2.9.0
pandas>=2.2.0
numpy>=1.26.0
python-multipart>=0.0.6
anthropic>=0.39.0
